add User component folder
