+++ src/components/User/UserLikedPosts.tsx
@@ -2,8 +2,8 @@ import { FC } from "react";
 import { trpc } from "@/utils/trpc";
 import useLikePost from "@/hooks/useLikePost";
 import useSavePost from "@/hooks/useSavePost";
-import PostSkeleton from "./Post/PostSkeleton";
-import Post from "./Post/Post";
+import PostSkeleton from "../Post/PostSkeleton";
+import Post from "../Post/Post";
 
 const UserLikedPosts: FC<{ user: string; isSelected: boolean }> = ({
   user,
