+++ src/components/User/UserSavedPosts.tsx
@@ -1,30 +1,29 @@
 import { FC } from "react";
 import { trpc } from "@/utils/trpc";
 import useLikePost from "@/hooks/useLikePost";
+import PostSkeleton from "../Post/PostSkeleton";
+import Post from "../Post/Post";
 import useSavePost from "@/hooks/useSavePost";
-import PostSkeleton from "./Post/PostSkeleton";
-import Post from "./Post/Post";
 
-const UserPosts: FC<{ user: string; isSelected: boolean }> = ({
+const UserSavedPosts: FC<{ user: string; isSelected: boolean }> = ({
   user,
   isSelected,
 }) => {
-  const postQuery = trpc.useQuery(["user.get-posts", { user }], {
+  const savedPostQuery = trpc.useQuery(["user.get-saved-posts", { user }], {
     enabled: isSelected,
   });
-
-  const { onLike, onUnlike } = useLikePost("user.get-posts", { user });
-  const { onSave, onUnsave } = useSavePost("user.get-posts", { user });
+  const { onLike, onUnlike } = useLikePost("user.get-saved-posts", { user });
+  const { onSave, onUnsave } = useSavePost("user.get-saved-posts", { user });
 
   return (
     <>
-      {postQuery.isLoading &&
+      {savedPostQuery.isLoading &&
         Array(13)
           .fill(0)
           .map((skeleton, idx) => <PostSkeleton key={idx} />)}
 
-      {postQuery?.data?.posts &&
-        postQuery.data.posts?.map((post) => (
+      {savedPostQuery?.data?.posts &&
+        savedPostQuery.data.posts?.map((post) => (
           <Post
             key={post.id}
             {...post}
@@ -34,7 +33,7 @@ const UserPosts: FC<{ user: string; isSelected: boolean }> = ({
             onUnsave={onUnsave}
           />
         ))}
-      {postQuery.data?.posts.length === 0 && (
+      {savedPostQuery.data?.posts.length === 0 && (
         <div className="flex flex-col items-center text-grayAlt">
           <p className="font-bold text-lg mt-6">Its empty here</p>
           <p className="text-xl">😢</p>
@@ -44,4 +43,4 @@ const UserPosts: FC<{ user: string; isSelected: boolean }> = ({
   );
 };
 
-export default UserPosts;
+export default UserSavedPosts;
